增加命令行工具 show

+++ commands/db.py
@@ -1,6 +1,6 @@
 import click
 
-from db.setup import clear_table, delete_row, drop_db, init_db
+from db.setup import clear_table, delete_row, drop_db, init_db, show_db
 
 
 @click.group(name="db")
@@ -13,6 +13,13 @@ def cli_db():
 def init():
     """初始化数据库表"""
     init_db()
+
+
+@cli_db.command("show")
+def show():
+    """展示数据库信息"""
+    doc = show_db()
+    click.secho(doc)
 
 
 @cli_db.command("drop")
@@ -29,13 +36,21 @@ def drop():
 @cli_db.command("clear")
 def clear():
     """清空数据表所有数据"""
-    count = clear_table()
-    click.secho(f"已删除所有数据表记录共{count}条", fg="green")
+    confirm = click.confirm("确定要删除所有数据表记录？此操作不可恢复！")
+    if not confirm:
+        click.secho("操作取消", fg="yellow")
+    else:
+        count = clear_table()
+        click.secho(f"已删除所有数据表记录共{count}条", fg="green")
 
 
 @cli_db.command("delete")
 @click.argument("days", type=click.IntRange(1, None))
 def delete(days):
-    """删除指定天数前的数据库记录"""
-    count = delete_row(days)
-    click.secho(f"已删除{days}天前的数据共{count}条", fg="green")
+    """删除指定天数前的数据库表记录"""
+    confirm = click.confirm(f"确定要删除{days}天前的所有数据表记录？此操作不可恢复！")
+    if not confirm:
+        click.secho("操作取消", fg="yellow")
+    else:
+        count = delete_row(days)
+        click.secho(f"已删除{days}天前的数据共{count}条", fg="green")
